- Refactorings in Chat related services. - RecipeDetail and Basket does not have error. - Localhost added to API instead of local ip
