+++ Presentation Tier/WebApp/Pages/Admin/AdminChat.razor
@@ -1,6 +1,6 @@
 ﻿@page "/adminchat"
 @using Microsoft.AspNetCore.SignalR.Client
-@using WebApp.Data.Hub
+@using WebApp.Data.HubConnection
 
 @inject NavigationManager navigationManager
