+++ Presentation Tier/WebApp/Pages/Customer/Basket.razor
@@ -1,6 +1,8 @@
 ﻿@page "/basket"
 @using WebApp.Data
 @using WebApp.Models
+@using MudBlazor
+@using global::Models
 
 @inject IBasketService BasketService
 @inject NavigationManager NavigationManager
@@ -18,27 +20,44 @@ else if (!recipesInBasket.Any())
 }
 else
 {
-    <MudTable Items="@recipesInBasket" Hover="true" Breakpoint="Breakpoint.Sm">
+    <MudTable Items="@recipes" Hover="true" Breakpoint="Breakpoint.Sm">
         <HeaderContent>
             <MudTh>Amount</MudTh>
             <MudTh>Name</MudTh>
             <MudTh>Price</MudTh>
+            <MudTh> </MudTh>
         </HeaderContent>
-        @foreach (var item in recipesInBasket)
-        {
-            <RowTemplate>
-                <MudTd DataLabel="Amount">@recipesInBasket.Count</MudTd>
-                <MudTd DataLabel="Name">@item.Amount</MudTd>
-                @* <MudTd DataLabel="Price">@item.Price</MudTd> *@
-            </RowTemplate>
-        }
+        <RowTemplate>
+            @foreach (var item in recipes)
+            {
+                <MudTd DataLabel="Amount">@item.Amount</MudTd>
+                <MudTd DataLabel="Name">@item.Name</MudTd>
+                <MudTd DataLabel="Price">@item.Price</MudTd>
+                <MudTd>
+                    <MudButton Variant="Variant.Filled" Color="Color.Secondary" @onclick="() => RemoveBasketItem(item)">Remove</MudButton>
+                </MudTd>
+            }
+        </RowTemplate>
+        <RowTemplate>
+            <MudTd DataLabel="Total amount">@totalAmount</MudTd>
+            <MudTd DataLabel="Total price">@totalPrice</MudTd>
+        </RowTemplate>
     </MudTable>
 }
 @code {
     private IList<BasketItem> recipesInBasket = new List<BasketItem>();
+    private IList<Recipe> recipes = new List<Recipe>();
+    private int? totalAmount { get; set; }
+    private double? totalPrice { get; set; }
 
     protected override async Task OnInitializedAsync()
     {
         recipesInBasket = await BasketService.GetAllBasketItems();
+        recipes = await BasketService.GetAllBasketRecipes();
+    }
+
+    private void RemoveBasketItem(Recipe recipe)
+    {
+        BasketService.RemoveRecipe(recipe);
     }
 }
