+++ Presentation Tier/WebApp/Pages/RecipeDetail.razor
@@ -1,11 +1,12 @@
 ﻿@page "/recipedetail/{id:int}"
 @using WebApp.Models
 @using WebApp.Data
+@using global::Models
 
 @inject IRecipeService RecipeService
 
 <h3>@recipeToShow.Name</h3>
-<h6>Type: @recipeToShow.Type</h6>
+<h6>Type: @recipeToShow.NutritionType</h6>
 <img src="@recipeToShow.Image" alt="">
 <p>Ingredients:</p>
 <ul>
@@ -25,10 +26,11 @@
 
     private Recipe recipeToShow = new Recipe();
     private IList<Ingredient> ingredientsInRecipe = new List<Ingredient>();
+    private IList<RecipeItem> recipeItems = new List<RecipeItem>();
 
     protected override async Task OnInitializedAsync()
     {
         recipeToShow = await RecipeService.GetRecipeAsync(id);
-        ingredientsInRecipe = recipeToShow.Ingredient;
+        recipeItems = recipeToShow.Ingredient;
     }
 }
